changed unsigned int to unsigned long long
started debug...
nodeNumber member is now a unsigned long long as unsigned int is
equivalant to unsigned __int32 and not unsigned __int64.
also added more files to gitignore **fucking github**

+++ project/project/BinaryTree.cpp
@@ -9,7 +9,7 @@
 * OUTPUT: none
 * RETURN VALUE: pointer to parent found or NULL if node doesn't exist
 *************************************************************************/
-treeNode * BinaryTree::findParent(int numberToDelete, treeNode * curr)
+treeNode * BinaryTree::findParent(unsigned long long numberToDelete, treeNode * curr)
 {
 	//first two if and else statements are redundant. Could probably lose them.
 	//May need to better define NULL base case for branch end points
@@ -50,7 +50,7 @@ bool BinaryTree::IsEmpty()
 * OUTPUT: none
 * RETURN VALUE: Returns a true/false (found/not found) boolean
 *************************************************************************/
-bool BinaryTree::FindNode(int numberToFind)
+bool BinaryTree::FindNode(unsigned long long numberToFind)
 {
 	treeNode* nodeFinder = root;
 	bool found = false,							//number has/has not been found (yet)
@@ -94,7 +94,7 @@ bool BinaryTree::FindNode(int numberToFind)
 * RETURN VALUE: returns the created node with initialized pointers to 
 * NULL, or NULL is the node could not be created.
 *************************************************************************/
-treeNode * BinaryTree::CreateNode(int numberToPlace, string steamID, GameList* topFive)
+treeNode * BinaryTree::CreateNode(unsigned long long numberToPlace, string steamID, GameList* topFive)
 {
 	treeNode * nodeCreated = new(nothrow) treeNode; 		//the 'nothrow' won't throw an exception if it fails, but will instead return NULL
 	if (nodeCreated == NULL)
@@ -120,7 +120,7 @@ treeNode * BinaryTree::CreateNode(int numberToPlace, string steamID, GameList* t
 * OUTPUT: none
 * RETURN VALUE: true if successful, otherwise false if duplicate exists
 *************************************************************************/
-bool BinaryTree::InsertNode(int id_64, string uName, GameList* gList)
+bool BinaryTree::InsertNode(unsigned long long id_64, string uName, GameList* gList)
 {
 	bool inserted = false;
 	//do a check for duplicates first
@@ -223,7 +223,7 @@ bool BinaryTree::InsertNode(treeNode * tempPtr)
 * OUTPUT: none
 * RETURN VALUE: none
 *************************************************************************/
-void BinaryTree::DeleteNode(int numberToDelete)
+void BinaryTree::DeleteNode(unsigned long long numberToDelete)
 {
 	//find first branch w/ null children and replace into deleted node
 	if (root->nodeNumber != numberToDelete) {
